VP-758 rename apiCaller to callApi for consistency
