+++ components/Member/__tests__/RegisterMemberSection.spec.js
@@ -5,7 +5,7 @@ import RegisterMemberSection from '../RegisterMemberSection'
 import { Provider } from 'react-redux'
 import reduxApi, { makeStore } from '../../../lib/redux/reduxApi'
 import adapterFetch from 'redux-api/lib/adapters/fetch'
-import { API_URL } from '../../../lib/apiCaller'
+import { API_URL } from '../../../lib/callApi'
 import fixture from './member.fixture.js'
 import { MemberStatus } from '../../../server/api/member/member.constants'
 import objectid from 'objectid'
