+++ components/Person/__tests__/PersonBadge.spec.js
@@ -2,7 +2,7 @@ import test from 'ava'
 import { PersonBadgeSection } from '../PersonBadge'
 import mockFetch from 'fetch-mock'
 import { mount } from 'enzyme'
-import { APP_URL } from '../../../lib/apiCaller'
+import { APP_URL } from '../../../lib/callApi'
 
 const badgeData = [{
   badgeclass: 'cqNQRgBaTVagim095Pldsw',
