+++ lib/__test__/callApi.spec.js
@@ -1,6 +1,6 @@
 import test from 'ava'
 import 'isomorphic-fetch'
-import callApi, { API_URL, convertCookiesToKvps } from '../apiCaller'
+import callApi, { API_URL, convertCookiesToKvps } from '../callApi'
 // import sinon from 'sinon'
 
 const { fetchMock } = require('fetch-mock')
