+++ lib/auth/__test__/auth.spec.js
@@ -4,7 +4,7 @@ import MemoryMongo from '../../../server/util/test-memory-mongo'
 import { getPersonFromUser, parseTokenToSession, parseUserToSession, createPersonFromUser, getUserFromServerCookie } from '../auth'
 import jwtDecode from 'jwt-decode'
 import fetchMock from 'fetch-mock'
-import { API_URL } from '../../../lib/apiCaller'
+import { API_URL } from '../../../lib/callApi'
 import people from '../../../server/api/person/__tests__/person.fixture'
 import { jwtData } from '../../../server/middleware/session/__tests__/setSession.fixture'
 import Person from '../../../server/api/person/person'
