+++ lib/auth/auth.js
@@ -1,6 +1,6 @@
 import jwtDecode from 'jwt-decode'
 import Cookie from 'js-cookie'
-import callApi from '../apiCaller'
+import callApi from '../callApi'
 
 export const parseTokenToSession = async (idToken) => {
   const user = jwtDecode(idToken)
