+++ lib/callApi.js
@@ -21,6 +21,7 @@ export const convertCookiesToKvps = (cookies) => {
 }
 
 const callApi = async (endpoint, method = 'get', body, cookies) => {
+  console.log('callApi', endpoint)
   const headers = {
     'content-type': 'application/json'
   }
@@ -36,11 +37,12 @@ const callApi = async (endpoint, method = 'get', body, cookies) => {
      */
     const requestBody = body === null ? undefined : body
     const response = await fetch(`${API_URL}/${endpoint}`, {
-      headers,
+      headers: body ? headers : undefined,
       method,
       body: JSON.stringify(requestBody)
     })
     if (!response.ok) {
+      console.error('CallAPI error response', response)
       // according to https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Promise/reject
       // it is useful to make reason an instanceof Error
       return Promise.reject(new Error(JSON.stringify({ status: response.status, statusText: response.statusText, message: response.body })))
